feat: protected goals routes

+++ backend/controllers/goalController.js
@@ -1,17 +1,12 @@
- const asyncHandler = require('express-async-handler') 
+const asyncHandler = require('express-async-handler') 
+const Goal = require('../models/goalModal')
+const User = require('../models/userModal')
 
- const Goal = require('../models/goalModal')
-// @desc Get Goals
-// @route GET /api/goals
-// @access Private
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({user: req.user.id})
     res.status(200).json(goals)
 })
 
-// @desc Set Goal
-// @route POST /api/goals
-// @access Private
 const setGoal = asyncHandler(async (req, res) => {
     if(!req.body.text){
         res.status(400)
@@ -20,14 +15,12 @@ const setGoal = asyncHandler(async (req, res) => {
 
     const goal = await Goal.create({
         text: req.body.text,
+        user: req.user.id
     })
 
     res.status(200).json(goal)
 })
 
-// @desc Update Goal
-// @route PUT /api/goals/:id
-// @access Private
 const updateGoal = asyncHandler(async (req, res) => {
     const goal = await Goal.findById(req.params.id)
 
@@ -36,20 +29,38 @@ const updateGoal = asyncHandler(async (req, res) => {
         throw new Error('Goal not found')
     }
 
-    const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new:true,})
+    const user = await User.findById(req.user.id)
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
 
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
+    const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new:true,})
     res.status(200).json(updatedGoal)
 })
 
-// @desc Delete Goal
-// @route DELETE /api/goals
-// @access Private
 const deleteGoal = asyncHandler(async (req, res) => {
     const goal = await Goal.findById(req.params.id)
 
     if(!goal){
         res.status(400)
         throw new Error('Goal not found')
+    }
+    
+    const user = await User.findById(req.user.id)
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     await goal.deleteOne()
